test(terminal-launch): move the portable Windows check to a sibling file
Keeps lib.rs near its budget and gives the platform-independent Windows
checks an obvious home, with a module doc explaining why they exist: the
cfg(not(unix)) tests beside them never run in this repo's Linux CI, which is
the blind spot that hid #715.

One baseline line moves (1216 -> 1219) for the module declaration itself.
