+++ crates/jcode-terminal-launch/src/lib.rs
@@ -777,6 +777,9 @@ fn macos_terminal_applescript(command: &TerminalCommand, cwd: &Path) -> String {
 }
 
 #[cfg(test)]
+#[path = "windows_portable_tests.rs"]
+mod windows_portable_tests;
+#[cfg(test)]
 mod tests {
     use super::*;
     use std::sync::Mutex;
@@ -1008,36 +1011,6 @@ mod tests {
         let shell = shell_command(&["jcode".to_string(), "it's ok".to_string()]);
         #[cfg(unix)]
         assert_eq!(shell, "'jcode' 'it'\"'\"'s ok'");
-    }
-
-    /// #715: the in-app resume spawn was a no-op off Unix, so the Windows
-    /// command line it *would* build was never exercised anywhere that runs.
-    ///
-    /// The existing Windows coverage is `cfg(not(unix))`, so it never runs in
-    /// this repo's Linux CI or on a maintainer's Linux box. `windows_command_line`
-    /// is compiled under `test` on every platform, so pin the resume invocation
-    /// here where it will actually be checked.
-    #[test]
-    fn windows_command_line_quotes_a_resume_invocation_on_every_platform() {
-        let line = windows_command_line(&[
-            r"C:\Program Files\jcode\jcode.exe".to_string(),
-            "--resume".to_string(),
-            "session_penguin_123".to_string(),
-        ]);
-
-        // The exe path contains a space, so it must be quoted or cmd.exe would
-        // treat "C:\Program" as the program.
-        assert!(
-            line.contains(r#""C:\Program Files\jcode\jcode.exe""#),
-            "unquoted program path would break on a default Windows install: {line}"
-        );
-        assert!(line.contains("--resume"), "{line}");
-        assert!(line.contains("session_penguin_123"), "{line}");
-        // Session ids are safe, so they must not be needlessly quoted.
-        assert!(
-            !line.contains(r#""--resume""#),
-            "quoting flags changes cmd.exe parsing: {line}"
-        );
     }
 
     #[test]
